[Refactor] Expose a higher-level API
We expect that our users will be more interested in machine-learning than in quantum computers,
at least at first, so we expose a new, higher-level, API, that hides most of the quantum details.
With this API, switching between the QutipEmulator, emu-mps or QPU is just a few lines of code
(well, one line of code + the connection details, username, password, project id).

This results in a tutorial that spends less time on the quantum aspects and more on the machine-learning.

Also, more tests.