[API] Rework the return of a compiler to avoid `async` and make `target` optional.
We don't want our users to need to understand `async`/`await` until they need to,
so we rework the output of extractors to make them usable without `async`/`await`.
Only users who write code for servers or interactive applications will need to `await`.

Also, we make `target` optional, because we want to be able to use our kernel with
data that does not come from the training material!

Note: breaking API changes, we'll need to bump version.