[Feature] Kernel integration - resolves #59 (#62)
Rewriting QuantumEvolutionKernel into FastQEK and IntegratedQEK. The latter that can transparently call the QPU or Emulator. This should make tutorials easier to write, but in production, this is not recommended, as this is more fragile (e.g. in case of server connection error), can be extremely slow (e.g. if there is a long waitline for the QPU) and blocks the main thread.